Move/rename existing code to "pre0"
