+++ .vscode/settings.json
@@ -5,5 +5,6 @@
   "eslint.packageManager": "yarn",
   "eslint.options": {
     "reportUnusedDisableDirectives": true
-  }
+  },
+  "jest.jestCommandLine": "yarn workspace @foxglove/mcap test"
 }
