+++ typescript/src/index.ts
@@ -1,4 +1,3 @@
-export { default as McapReader } from "./McapReader";
-export * from "./parse";
-export * from "./types";
-export { default as McapWriter } from "./McapWriter";
+export { default as McapPre0Reader } from "./pre0/McapPre0Reader";
+export * as Pre0Types from "./pre0/types";
+export { default as McapPre0Writer } from "./pre0/McapPre0Writer";
