+++ typescript/src/pre0/constants.ts
